Fixes to mocking for dg-download tests #86
