+++ packages/datagateway-download/__mocks__/axios.ts
@@ -1,3 +1,5 @@
+import axios from "axios";
+
 export default {
   get: jest.fn(path => {
     if (path.includes('/topcat/user/cart/')) {
@@ -7,4 +9,5 @@ export default {
     }
   }),
   post: jest.fn(() => Promise.resolve({ data: {} })),
+  CancelToken: axios.CancelToken,
 };
